refactor(backend): 🚚 rename module
This makes the inherent factory pattern clearer
