+++ app/backend/app.ts
@@ -1,5 +1,5 @@
 import { serveStatic } from "hono/bun";
-import { createApi, createRouter } from "#lib/create-app";
+import { createApi, createRouter } from "#lib/factory";
 import { chat } from "./routes/chat";
 
 // * API
